Fix inspector warnings introduced by recent mergences

+++ codec-socks/src/main/java/io/netty/codec/socks/SocksAuthRequestDecoder.java
@@ -45,7 +45,7 @@ public class SocksAuthRequestDecoder extends ReplayingDecoder<SocksRequest, Sock
     public SocksRequest decode(ChannelHandlerContext ctx, ByteBuf byteBuf) throws Exception {
         switch (state()) {
             case CHECK_PROTOCOL_VERSION: {
-                version = SocksMessage.ProtocolVersion.fromByte((byte) byteBuf.readByte());
+                version = SocksMessage.ProtocolVersion.fromByte(byteBuf.readByte());
                 if (version != SocksMessage.ProtocolVersion.SOCKS5) {
                     break;
                 }
